requirements.txt
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
gunicorn==19.7.1
idna==2.6
pew==1.1.2
pipenv==10.1.0
psutil==5.3.1
pygame==1.9.3
python-telegram-bot==9.0.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
